refactor(runs): remove legacy agent orchestration path (AYC-769)
