+++ ayunis-core-backend/.env.example
@@ -42,11 +42,9 @@ APPSIGNAL_PUSH_API_KEY=
 # on the Node side first and surface as intermittent 502s.
 HTTP_KEEP_ALIVE_TIMEOUT_MS=
 
-# Agent runtime rollout (default false). Routes chat runs through
-# @ayunis/agent-runtime when true; enable in staging, then internal, then
-# production. Verify the effective value after deployment or rollback via
-# GET /api/feature-toggles. See ../docs/runbooks/agent-runtime-rollout.md.
-FEATURE_AGENT_RUNTIME_ENABLED=false
+# Employee-facing SSO login rollout. Superadmin SSO configuration remains
+# available when disabled.
+FEATURE_SSO_LOGIN_ENABLED=false
 
 # Frontend
 # Including 'https://', used for sending invite emails, password reset mails, etc.
