+++ ayunis-core-backend/src/app/SUMMARY.md
@@ -3,4 +3,4 @@ Root NestJS module wiring all domain modules and infrastructure together.
 
 The app module is the application entry point. It imports and configures all domain modules (models, agents, threads, runs, messages, tools, prompts, sources, RAG, retrievers, storage, MCP, usage, transcriptions, shares, knowledge-bases, artifacts, letterheads), IAM, scheduling, static file serving, CLS transactional support, structured Pino logging, metrics, and all configuration modules. It also provides the health check and cloud-detection endpoints.
 
-The `AppModule` is the root NestJS module that composes the entire application. It loads all configuration via `ConfigModule.forRoot()` with numerous config factories (models, auth, typeorm, embeddings, storage, web, app, admin, legal, subscriptions, emails, internet-search, MCP, tools). Database connectivity uses `TypeOrmModule.forRootAsync()` with the shared `dataSource`, including a vector type hack for pgvector support. Request-scoped context is initialized via `ClsModule.forRoot()` with `ClsPluginTransactional` for TypeORM transaction propagation. `LoggerModule` installs the shared Pino configuration; bootstrap buffers framework logs until nestjs-pino is ready and then uses it as Nest's framework logger. `ScheduleModule` is included for cron jobs; AppSignal monitoring is bootstrapped outside the module graph via `--require ./appsignal.cjs`; after `app.listen` succeeds, bootstrap starts the continuous `backend_process` process-monitor heartbeat. Public `AppController` endpoints include `GET /` for deployment/registration status, `GET /health`, and `GET /feature-toggles`, whose `agentRuntimeEnabled` value lets operators verify the effective rollout configuration after deployment or rollback. Middleware (`CookieParserMiddleware`, `SecurityHeadersMiddleware`) is applied globally. `ServeStaticModule` serves the frontend build from the `client` directory in production.
+The `AppModule` is the root NestJS module that composes the entire application. It loads all configuration via `ConfigModule.forRoot()` with numerous config factories (models, auth, typeorm, embeddings, storage, web, app, admin, legal, subscriptions, emails, internet-search, MCP, tools). Database connectivity uses `TypeOrmModule.forRootAsync()` with the shared `dataSource`, including a vector type hack for pgvector support. Request-scoped context is initialized via `ClsModule.forRoot()` with `ClsPluginTransactional` for TypeORM transaction propagation. `LoggerModule` installs the shared Pino configuration; bootstrap buffers framework logs until nestjs-pino is ready and then uses it as Nest's framework logger. `ScheduleModule` is included for cron jobs; AppSignal monitoring is bootstrapped outside the module graph via `--require ./appsignal.cjs`; after `app.listen` succeeds, bootstrap starts the continuous `backend_process` process-monitor heartbeat. Public `AppController` endpoints include `GET /` for deployment/registration status, `GET /health`, and `GET /feature-toggles` for the current user-facing feature toggle states. Middleware (`CookieParserMiddleware`, `SecurityHeadersMiddleware`) is applied globally. `ServeStaticModule` serves the frontend build from the `client` directory in production.
