+++ ayunis-core-backend/src/config/SUMMARY.md
@@ -7,6 +7,6 @@ The config directory contains two categories of files:
 
 **NestJS registerAs Factories** — The `*.config.ts` files are NestJS `registerAs` configuration factories that map environment variables to typed configuration objects. Each file covers a specific concern: admin tokens, app settings, authentication/JWT, emails/SMTP, embeddings, evals/Langfuse, feature toggles, internet search/Brave, legal versioning, Locaboo, MCP URLs, model provider API keys, municipal SSO OIDC, retrieval/Gotenberg, storage/MinIO, subscriptions/pricing, tools limits, TypeORM/Postgres, URL retriever limits, and web/CORS. `frontend.config.ts` whitelists the runtime values exposed by `/config.js`, including optional GeoJSON-widget basemap tile URL and attribution overrides.
 
-The configuration module provides strongly-typed, namespaced configuration for every backend subsystem via NestJS `registerAs` factories. `app.config.ts` controls port, deployment mode (cloud/self-hosted), registration toggle, frontend URLs, webhook endpoints (`ORG_EVENTS_WEBHOOK_URL`, disabled by `DISABLE_ORG_EVENTS_WEBHOOKS=true`), the HMAC-SHA256 outbound webhook signing secret (`WEBHOOK_SIGNING_SECRET`), and deterministic mock inference for test/e2e stacks (`MOCK_INFERENCE`). `authentication.config.ts` configures JWT secrets, token expiration for access/refresh/invite/email-confirmation/password-reset tokens, cookie settings, and auth provider selection (local vs cloud). `sso-oidc.config.ts` holds the optional-as-a-group Zitadel issuer, confidential client credentials, exact callback URL, and SSO reauthentication window for the municipal SSO relying party. `models.config.ts` holds API keys and base URLs for all LLM providers (Mistral, OpenAI, Anthropic, Bedrock, Ollama, Synaforce, Ayunis, OTC, Gemini, Azure). `typeorm.config.ts` builds PostgreSQL `DataSourceOptions` with environment-specific entity/migration paths and SSL settings for production. `storage.config.ts` defines MinIO connection parameters. `subscriptions.config.ts` sets per-seat pricing and trial message limits. `emails.config.ts` maps SMTP host/port/credentials. `legal.config.ts` tracks versioned terms-of-service and provider-specific legal acceptances. `features.config.ts` defines feature toggles (`FeatureFlag` enum) for agents, knowledge bases, prompts, skills, and the agent-runtime execution path, each controlled by a `FEATURE_*_ENABLED` environment variable with sensible defaults. `url.config.ts` (the `url` namespace) bounds the URL retriever's outbound fetches: `URL_RETRIEVER_TIMEOUT_MS` (fetch timeout, default 5s) and `URL_RETRIEVER_MAX_DOWNLOAD_BYTES` (maximum bytes buffered from a fetched URL, default 25 MB). The root-level factories are collected in `root-configs.ts` (`rootConfigs`), which every Nest entrypoint (`AppModule`, `CliModule`) passes to `ConfigModule.forRoot({ load: rootConfigs, validate: validateEnv })` — entrypoints must load the shared list rather than hand-picking namespaces, because their transitive module graphs read namespaced keys (e.g. `auth.jwt.secret` via `getOrThrow`) that fail at bootstrap if the namespace is missing. Feature-scoped factories (e.g. `retrieval.config.ts`) are loaded by their own modules via `ConfigModule.forFeature`.
+The configuration module provides strongly-typed, namespaced configuration for every backend subsystem via NestJS `registerAs` factories. `app.config.ts` controls port, deployment mode (cloud/self-hosted), registration toggle, frontend URLs, webhook endpoints (`ORG_EVENTS_WEBHOOK_URL`, disabled by `DISABLE_ORG_EVENTS_WEBHOOKS=true`), the HMAC-SHA256 outbound webhook signing secret (`WEBHOOK_SIGNING_SECRET`), and deterministic mock inference for test/e2e stacks (`MOCK_INFERENCE`). `authentication.config.ts` configures JWT secrets, token expiration for access/refresh/invite/email-confirmation/password-reset tokens, cookie settings, and auth provider selection (local vs cloud). `sso-oidc.config.ts` holds the optional-as-a-group Zitadel issuer, confidential client credentials, exact callback URL, and SSO reauthentication window for the municipal SSO relying party. `models.config.ts` holds API keys and base URLs for all LLM providers (Mistral, OpenAI, Anthropic, Bedrock, Ollama, Synaforce, Ayunis, OTC, Gemini, Azure). `typeorm.config.ts` builds PostgreSQL `DataSourceOptions` with environment-specific entity/migration paths and SSL settings for production. `storage.config.ts` defines MinIO connection parameters. `subscriptions.config.ts` sets per-seat pricing and trial message limits. `emails.config.ts` maps SMTP host/port/credentials. `legal.config.ts` tracks versioned terms-of-service and provider-specific legal acceptances. `features.config.ts` defines feature toggles (`FeatureFlag` enum) for knowledge bases, letterheads, skills, workspaces, and employee-facing SSO login, each controlled by a `FEATURE_*_ENABLED` environment variable with sensible defaults. `url.config.ts` (the `url` namespace) bounds the URL retriever's outbound fetches: `URL_RETRIEVER_TIMEOUT_MS` (fetch timeout, default 5s) and `URL_RETRIEVER_MAX_DOWNLOAD_BYTES` (maximum bytes buffered from a fetched URL, default 25 MB). The root-level factories are collected in `root-configs.ts` (`rootConfigs`), which every Nest entrypoint (`AppModule`, `CliModule`) passes to `ConfigModule.forRoot({ load: rootConfigs, validate: validateEnv })` — entrypoints must load the shared list rather than hand-picking namespaces, because their transitive module graphs read namespaced keys (e.g. `auth.jwt.secret` via `getOrThrow`) that fail at bootstrap if the namespace is missing. Feature-scoped factories (e.g. `retrieval.config.ts`) are loaded by their own modules via `ConfigModule.forFeature`.
 
 **Boot-time env validation** — `env.validation.ts` is the single source of truth for a *valid* environment (it mirrors `.env.example`). Its `validateEnv` function runs via the `ConfigModule.forRoot({ validate })` hook for every entrypoint that loads `rootConfigs`, so a missing/invalid variable fails startup with one aggregated message listing **every** problem at once — instead of silently falling back to a default (which can mask a typo'd var) or crashing lazily deep in the DI graph. It validates required secrets (`JWT_SECRET`, `COOKIE_SECRET` in every environment; no insecure default fallback), production-only rules (`COOKIE_SECURE=true`, MinIO credentials, `REDIS_PASSWORD`), and the shape of optional variables *when present* (integers, booleans, enums). Municipal SSO is optional, but `SSO_OIDC_ISSUER`, `SSO_OIDC_CLIENT_ID`, `SSO_OIDC_CLIENT_SECRET`, `SSO_OIDC_CALLBACK_URL`, and the app-only `SSO_LOGIN_TRANSACTION_ENCRYPTION_KEY` must be supplied together; issuer and callback URLs require HTTPS except for loopback development. These requiredness rules were consolidated here from the former per-factory asserts (auth, storage, redis); the factories now only read values, while their default fallbacks stay in the factory. Validation uses `class-validator`/`class-transformer` (already project dependencies). Note: `WEBHOOK_SIGNING_SECRET` is validated separately in `webhooks.module.ts` `OnModuleInit` because it warns (rather than throws) outside production.
