+++ ayunis-core-backend/src/config/env.validation.ts
@@ -192,7 +192,7 @@ export class EnvironmentVariables {
   @IsOptional() @IsIn(BOOLEAN_STRINGS) FEATURE_LETTERHEADS_ENABLED?: string;
   @IsOptional() @IsIn(BOOLEAN_STRINGS) FEATURE_SKILLS_ENABLED?: string;
   @IsOptional() @IsIn(BOOLEAN_STRINGS) FEATURE_WORKSPACES_ENABLED?: string;
-  @IsOptional() @IsIn(BOOLEAN_STRINGS) FEATURE_AGENT_RUNTIME_ENABLED?: string;
+  @IsOptional() @IsIn(BOOLEAN_STRINGS) FEATURE_SSO_LOGIN_ENABLED?: string;
 
   // Retention
   @IsOptional() @IsIn(BOOLEAN_STRINGS) RETENTION_DRY_RUN?: string;
