+++ ayunis-core-backend/src/domain/messages/SUMMARY.md
@@ -3,7 +3,7 @@ Thread messages with typed roles and rich content
 
 Messages represent individual conversation entries within threads, supporting user, assistant, system, and tool-result roles. Each message contains typed content blocks including text, images, thinking, tool-use, and tool-result entries.
 
-The messages module handles storage and creation of conversation messages within threads. The abstract `Message` entity specializes into `UserMessage`, `AssistantMessage`, `SystemMessage`, and `ToolResultMessage`, each containing polymorphic `MessageContent` blocks (text, image, thinking, tool-use, tool-result). Tool-use content may carry a transient stream envelope for SSE presentation of accumulated raw arguments or an invalid terminal call; that envelope is not part of persisted conversation history. Key use cases include creating messages by role, saving streamed assistant responses, deleting messages, counting tokens, trimming messages for context windows, and cleaning up orphaned images. Final assistant-message and tool-result persistence treat the messages-to-thread foreign-key violation as an expected no-op when a thread is deleted during inference; the shared unexpected-use-case boundary consistently wraps every other tool-result persistence failure as `MESSAGE_CREATION_FAILED`. The `ImageContentService` manages image attachments in messages.
+The messages module handles storage and creation of conversation messages within threads. The abstract `Message` entity specializes into `UserMessage`, `AssistantMessage`, `SystemMessage`, and `ToolResultMessage`, each containing polymorphic `MessageContent` blocks (text, image, thinking, tool-use, tool-result). Tool-use content may carry a transient stream envelope for SSE presentation of accumulated raw arguments or an invalid terminal call; that envelope is not part of persisted conversation history. Key use cases create user, system, and tool-result messages, save streamed assistant responses, delete messages, count tokens, and clean up orphaned images. The runs module selects complete turns for the model context through its runtime history materializer. Final assistant-message and tool-result persistence treat the messages-to-thread foreign-key violation as an expected no-op when a thread is deleted during inference; the shared unexpected-use-case boundary consistently wraps every other tool-result persistence failure as `MESSAGE_CREATION_FAILED`. The `ImageContentService` manages image attachments in messages.
 
 Domain events are defined in `application/events/`: `UserMessageCreatedEvent` (emitted when a user sends a message, includes userId, orgId, threadId, messageId) and `AssistantMessageCreatedEvent` (emitted when an assistant message is persisted, includes userId, orgId, threadId, messageId). Each event class exposes a static `EVENT_NAME` string for type-safe listener registration.
