--- ayunis-core-backend/src/domain/messages/application/use-cases/trim-messages-for-context/trim-messages-for-context.command.ts
@@ -1,10 +1,0 @@
-import type { TokenCounterType } from 'src/common/token-counter/application/ports/token-counter.handler.port';
-import type { Message } from 'src/domain/messages/domain/message.entity';
-
-export class TrimMessagesForContextCommand {
-  constructor(
-    public readonly messages: Message[],
-    public readonly maxTokens: number,
-    public readonly counterType?: TokenCounterType,
-  ) {}
-}
