+++ ayunis-core-backend/src/domain/models/SUMMARY.md
@@ -72,9 +72,9 @@ The models module is the central registry for AI model configuration. The abstra
 - **MapMessagesToInferenceUseCase** (`application/use-cases/map-messages-to-inference`): Converts backend thread messages into provider-agnostic inference messages for consumers outside the models module, applying the same `stripReplayedToolNulls` history sanitation before mapping.
 - **GenerateImageUseCase** (`application/use-cases/generate-image/generate-image.use-case.ts`): Resolves the correct image-generation handler via `ImageGenerationHandlerRegistry` and executes the generation request. The returned `ImageGenerationResult` carries an optional `usage: { inputTokens, outputTokens, totalTokens }` block populated by the Azure handler when the model reports token usage (e.g. `gpt-image-1`); callers forward it to `CollectUsageAsyncService` for credit accounting. `ImageGenerationModel` carries the same `inputTokenCost` / `outputTokenCost` (EUR per million tokens) fields as `LanguageModel`, so cost calculation reuses the language-model formula in `CollectUsageUseCase`.
 
-The two inference use-cases (`GetInferenceUseCase`, `StreamInferenceUseCase`) are the error choke points for provider SDK failures on the legacy path (the per-provider adapter files are thin credential wrappers): `ApplicationError` passes through, client aborts become `InferenceAbortedError` (499, streaming only), transport failures and upstream 5xx become the shared `ProviderUnavailableError` family via `wrapProviderFailure` from `src/common/errors` (stable AppSignal grouping per provider+failure-class, AYC-538), and everything else — including upstream 4xx, which may be our own request-building bug — becomes `InferenceFailedError` (500). Completed streams with unparseable tool-call arguments are classified downstream as `InferenceMalformedToolCallError`, separately from both generic failures and `InferenceTokenLimitError`, so the runs module can safely retry before durable output (AYC-741). Upstream request rejections also emit privacy-safe structured diagnostics (provider code/type, parameter path, request ID, and an allowlisted reason category) without persisting raw SDK messages that may echo prompts or tool payloads. Runtime-backed stream handlers retain ownership of provider chunk transforms; `ResolveModelProviderUseCase` returns the transformed provider for the runs module to decorate with host-side error, cancellation, timeout, and metrics behavior. Azure streaming disables nested OpenAI SDK retries and uses the host boundaries' explicit pre-output attempt budgets instead, keeping cancellation, diagnostics, and total request counts observable.
+The two direct inference use-cases (`GetInferenceUseCase`, `StreamInferenceUseCase`) are the error choke points for callers such as OpenAI compatibility, title generation, and personalized system prompts (the per-provider adapter files are thin credential wrappers): `ApplicationError` passes through, client aborts become `InferenceAbortedError` (499, streaming only), transport failures and upstream 5xx become the shared `ProviderUnavailableError` family via `wrapProviderFailure` from `src/common/errors` (stable AppSignal grouping per provider+failure-class, AYC-538), and everything else — including upstream 4xx, which may be our own request-building bug — becomes `InferenceFailedError` (500). Upstream request rejections also emit privacy-safe structured diagnostics (provider code/type, parameter path, request ID, and an allowlisted reason category) without persisting raw SDK messages that may echo prompts or tool payloads. Runtime-backed stream handlers retain ownership of provider chunk transforms; `ResolveModelProviderUseCase` returns the transformed provider for the runs module to decorate with host-side error, cancellation, timeout, and metrics behavior. Azure streaming disables nested OpenAI SDK retries and uses the host boundaries' explicit pre-output attempt budgets instead, keeping cancellation, diagnostics, and total request counts observable.
 
-Legacy and SDK runtime streams use the same `StreamIdleWatchdog` from `src/common/streaming/stream-idle-watchdog.ts`. Its 3-minute timer (deliberately generous — reasoning models stream nothing while thinking, and provider keep-alive pings never reach the host, AYC-665) starts only after the first provider chunk and resets after every later chunk. Provider SDK request timeouts remain responsible for time-to-first-byte.
+Direct-inference and agent-runtime streams use the same `StreamIdleWatchdog` from `src/common/streaming/stream-idle-watchdog.ts`. Its 3-minute timer (deliberately generous — reasoning models stream nothing while thinking, and provider keep-alive pings never reach the host, AYC-665) starts only after the first provider chunk and resets after every later chunk. Provider SDK request timeouts remain responsible for time-to-first-byte.
 
 ## Infrastructure Services
