+++ ayunis-core-backend/src/domain/runs/application/use-cases/execute-run/execute-run.types.ts
@@ -4,15 +4,15 @@ import type { LanguageModel } from 'src/domain/models/domain/models/language.mod
 import type { Skill } from 'src/domain/skills/domain/skill.entity';
 import type { Thread } from 'src/domain/threads/domain/thread.entity';
 import type { Tool as BackendTool } from 'src/domain/tools/domain/tool.entity';
-import type { RuntimeToolIntegrationRegistry } from '../../agent-runtime/runtime-tool-integration.registry';
+import type { RuntimeToolIntegrationRegistry } from 'src/domain/runs/application/agent-runtime/runtime-tool-integration.registry';
 
-export interface PreparedRuntimeTools {
+export interface PreparedTools {
   tools: RuntimeTool[];
   backendTools: BackendTool[];
   toolIntegrations: RuntimeToolIntegrationRegistry;
 }
 
-export interface PreparedRuntimeRun extends PreparedRuntimeTools {
+export interface PreparedRun extends PreparedTools {
   thread: Thread;
   model: LanguageModel;
   orgId: UUID;
@@ -20,7 +20,6 @@ export interface PreparedRuntimeRun extends PreparedRuntimeTools {
   isAnonymous: boolean;
   instructions: string;
   activeSkills: Skill[];
-  canUseTools: boolean;
   skillInstructions?: string;
   activatedSkillName?: string;
 }
