--- ayunis-core-backend/src/domain/runs/application/use-cases/execute-run/run-params.interface.ts
@@ -1,26 +1,0 @@
-import type { UUID } from 'crypto';
-import type { Thread } from 'src/domain/threads/domain/thread.entity';
-import type { Tool } from 'src/domain/tools/domain/tool.entity';
-import type { LanguageModel } from 'src/domain/models/domain/models/language.model';
-import type {
-  RunUserInput,
-  RunToolResultInput,
-} from 'src/domain/runs/domain/run-input.entity';
-import type { Skill } from 'src/domain/skills/domain/skill.entity';
-import type { WorkspaceRunContext } from 'src/domain/workspaces/domain/workspace-run-context.entity';
-
-export interface RunParams {
-  thread: Thread;
-  tools: Tool[];
-  model: LanguageModel;
-  input: RunUserInput | RunToolResultInput;
-  instructions?: string;
-  streaming?: boolean;
-  orgId: UUID;
-  isAnonymous: boolean;
-  activeSkills: Skill[];
-  workspaceContext?: WorkspaceRunContext;
-  skillId?: UUID;
-  skillInstructions?: string;
-  activatedSkillName?: string;
-}
