+++ ayunis-core-backend/src/domain/usage/SUMMARY.md
@@ -7,7 +7,7 @@ The usage module provides comprehensive analytics for AI resource consumption. T
 
 **Organization-scoped use cases**: collecting usage after each run (`CollectUsageUseCase` — includes `calculateCredits` logic that converts cost to credits via `GetCreditsPerEuroUseCase` from **platform-config**; accepts either a `LanguageModel` or an `ImageGenerationModel` via `CollectableUsageModel`; emits a `UsageCollectedEvent` domain event after persisting, enabling downstream listeners such as webhook dispatchers; webhook payloads preserve the user/API-key principal distinction), querying provider-level usage (`GetProviderUsageUseCase`), model distribution analytics (`GetModelDistributionUseCase`), aggregate stats (`GetUsageStatsUseCase`), per-user usage reports (`GetUserUsageUseCase`), and monthly credit usage tracking (`GetMonthlyCreditUsageUseCase`).
 
-**Services**: `CollectUsageAsyncService` (`application/services/collect-usage-async.service.ts`) is the fire-and-forget wrapper around `CollectUsageUseCase`. It invokes `CollectUsageUseCase` and emits `TokensConsumedEvent` (`application/events/tokens-consumed.event.ts`) **after the usage row is persisted**, so listeners can read their own writes (metrics, `iam/budget-alerts` budget evaluation); a failed persist emits no token event. Run callers may also supply the explicit `legacy` or `agent_runtime` path, producing a low-cardinality `RunUsageCollectionEvent` on persistence success or failure; non-run callers omit the path. Errors are logged but do not block the calling flow. Exported from `UsageModule` so both the runs module (streaming/non-streaming inference) and the tools module (`GenerateImageToolHandler`) can drive collection.
+**Services**: `CollectUsageAsyncService` (`application/services/collect-usage-async.service.ts`) is the fire-and-forget wrapper around `CollectUsageUseCase`. It invokes `CollectUsageUseCase` and emits `TokensConsumedEvent` (`application/events/tokens-consumed.event.ts`) **after the usage row is persisted**, so listeners can read their own writes (metrics, `iam/budget-alerts` budget evaluation); a failed persist emits no token event. The agent-runtime usage hook supplies the `agent_runtime` path, producing a low-cardinality `RunUsageCollectionEvent` on persistence success or failure; the `legacy` type value remains temporarily for rollout-era monitoring continuity, while non-run callers omit the path. Errors are logged but do not block the calling flow. Exported from `UsageModule` so the runs module and tools such as `GenerateImageToolHandler` can drive collection.
 
 **Credit-based aggregations**: provider usage, model distribution, and aggregate stats all sum `creditsConsumed` (preserving fractional precision via `Number()`), not raw tokens. The raw `inputTokens`/`outputTokens`/`totalTokens` columns remain on `UsageRecord` (written by `CollectUsageUseCase`) but are no longer surfaced by the read/analytics views. The credit-typed read fields are `UsageStats.totalCredits`, `ProviderUsage.credits`, `ModelDistribution.credits`, and `TimeSeriesPoint.credits`. Monthly scoped reads include single and batched API-key aggregations; the batched path uses one set-based `IN`/`GROUP BY apiKeyId` query so credit-limit overviews do not issue an N+1 query.
